fix: update expected blocks for failing tests
The real issue is using callPublicFn instead of mineBlock, as it does not provide the block height the TX was sent in. A more sustainable solution would be to use mineBlock with the tx data, then using its height property to perform the check. This error will get triggered again if anything is changed around the deployment plan, e.g. a new contract is added.